♻️(admin) refactor admins, set fields read-only
