+++ src/tycho/infrastructure/django_apps/candidate/admin.py
@@ -5,12 +5,11 @@ from django.contrib import admin
 from infrastructure.django_apps.candidate.models.cv_metadata import CVMetadataModel
 
 
+@admin.register(CVMetadataModel)
 class CVMetadataAdmin(admin.ModelAdmin):
     """Admin configuration for CVMetadata model."""
 
-    list_display = ("filename", "id", "created_at", "extracted_text", "search_query")
+    list_display = ("filename", "id", "created_at", "search_query")
     list_filter = ("created_at",)
     search_fields = ("filename", "id")
-
-
-admin.site.register(CVMetadataModel, CVMetadataAdmin)
+    readonly_fields = [f.name for f in CVMetadataModel._meta.get_fields()]
