+++ src/tycho/infrastructure/django_apps/ingestion/admin.py
@@ -2,59 +2,10 @@
 
 from django.contrib import admin
 
-from infrastructure.django_apps.ingestion.models import RawDocument
-from infrastructure.django_apps.shared.models import vectorized_document
-from infrastructure.django_apps.shared.models.concours import ConcoursModel
-from infrastructure.django_apps.shared.models.corps import CorpsModel
-from infrastructure.django_apps.shared.models.offer import OfferModel
+from infrastructure.django_apps.ingestion.models.raw_document import RawDocument
 
 
-class OfferAdmin(admin.ModelAdmin):
-    """Admin interface for Offer model (read-only)."""
-
-    list_display = (
-        "id",
-        "external_id",
-        "verse",
-        "title",
-        "category",
-        "contract_type",
-        "region",
-        "department",
-        "beginning_date",
-        "created_at",
-        "updated_at",
-    )
-    list_filter = (
-        "verse",
-        "category",
-        "contract_type",
-        "region",
-        "created_at",
-        "updated_at",
-    )
-    search_fields = ("external_id", "title", "profile", "mission", "organization")
-    readonly_fields = (
-        "id",
-        "external_id",
-        "verse",
-        "title",
-        "profile",
-        "mission",
-        "category",
-        "contract_type",
-        "organization",
-        "offer_url",
-        "country",
-        "region",
-        "department",
-        "publication_date",
-        "beginning_date",
-        "created_at",
-        "updated_at",
-    )
-
-
+@admin.register(RawDocument)
 class RawDocumentAdmin(admin.ModelAdmin):
     """Admin interface for RawDocument model."""
 
@@ -62,107 +13,8 @@ class RawDocumentAdmin(admin.ModelAdmin):
         "id",
         "external_id",
         "document_type",
-        "raw_data",
         "created_at",
         "updated_at",
     )
     list_filter = ("document_type", "created_at", "updated_at")
-
-
-class CorpsAdmin(admin.ModelAdmin):
-    """Admin interface for Corps model."""
-
-    list_display = (
-        "id",
-        "code",
-        "short_label",
-        "long_label",
-        "category",
-        "ministry",
-        "diploma_level",
-        "access_modalities",
-    )
-    list_filter = ("category", "ministry", "diploma_level")
-    search_fields = ("code", "short_label", "long_label")
-    readonly_fields = ("id", "created_at", "updated_at")
-
-    fieldsets = (
-        (
-            "Informations générales",
-            {"fields": ("id", "code", "short_label", "long_label")},
-        ),
-        ("Classification", {"fields": ("category", "ministry", "diploma_level")}),
-        ("Modalités d'accès", {"fields": ("access_modalities",)}),
-        (
-            "Métadonnées",
-            {"fields": ("created_at", "updated_at"), "classes": ("collapse",)},
-        ),
-    )
-
-
-class ConcoursAdmin(admin.ModelAdmin):
-    """Admin interface for Concours model."""
-
-    list_display = (
-        "id",
-        "nor_original",
-        "corps",
-        "grade",
-        "category",
-        "ministry",
-        "open_position_number",
-        "written_exam_date",
-    )
-    list_filter = ("category", "ministry", "written_exam_date")
-    search_fields = ("nor_original", "corps", "grade")
-    readonly_fields = ("id", "created_at", "updated_at")
-
-    fieldsets = (
-        (
-            "Informations générales",
-            {"fields": ("id", "nor_original", "corps", "grade")},
-        ),
-        ("Classification", {"fields": ("category", "ministry")}),
-        ("Modalités d'accès", {"fields": ("access_modality",)}),
-        (
-            "Détails du concours",
-            {"fields": ("open_position_number", "written_exam_date")},
-        ),
-        ("NOR associés", {"fields": ("nor_list",)}),
-        (
-            "Métadonnées",
-            {"fields": ("created_at", "updated_at"), "classes": ("collapse",)},
-        ),
-    )
-
-
-class VectorizedDocumentAdmin(admin.ModelAdmin):
-    """Admin interface for VectorizedDocument model."""
-
-    list_display = (
-        "id",
-        "document_id",
-        "document_type",
-        "content",
-        "metadata",
-        "created_at",
-        "updated_at",
-    )
-    list_filter = ("document_type", "created_at", "updated_at")
-    search_fields = ("document_id", "content")
-    readonly_fields = (
-        "id",
-        "content",
-        "metadata",
-        "created_at",
-        "updated_at",
-    )
-
-
-admin.site.register(RawDocument, RawDocumentAdmin)
-admin.site.register(OfferModel, OfferAdmin)
-admin.site.register(CorpsModel, CorpsAdmin)
-admin.site.register(ConcoursModel, ConcoursAdmin)
-admin.site.register(
-    vectorized_document.VectorizedDocumentModel, VectorizedDocumentAdmin
-)
+    readonly_fields = [f.name for f in RawDocument._meta.get_fields()]
